refactor(serialization): move serialization classes to platform specific projects.
