+++ HateoasNet.Core/Serialization/GuidConverter.cs
@@ -2,7 +2,7 @@
 using System.Text.Json;
 using System.Text.Json.Serialization;
 
-namespace HateoasNet.Converters
+namespace HateoasNet.Core.Serialization
 {
 	public class GuidConverter : JsonConverter<Guid>
 	{
